pyproject.toml
[tool.poetry]
name = "async-langgraph-demo"
version = "0.1.0"
description = ""
authors = ["Brais Maneiro Sánchez <brais.maneiro@netexlearning.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.2.53"
python-dotenv = "^1.0.1"
langgraph-checkpoint-sqlite = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
